backend/requirements.txt
flask==2.3.3
python-docx==1.0.1
werkzeug==2.3.7
python-dotenv==1.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
docxtpl==0.16.7
lxml>=5.0.0
flasgger==0.9.7.1
psutil>=5.9.0
flask-socketio>=5.3.0
eventlet>=0.33.0
celery>=5.3.0
redis>=5.0.0
mammoth>=1.6.0
# Commercial / Auth Stack
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-jwt-extended>=4.5.3
psycopg2-binary>=2.9.9
email-validator>=2.1.0
stripe>=7.0.0
# Phase 1: Authentication & Security
sendgrid>=6.10.0
pyotp>=2.9.0
qrcode>=7.4.2
authlib>=1.3.0
pillow>=10.0.0
